FIXED : several issues fiexed
- 자신의 글이 아니여도 수정, 삭제되는 문제.
- 로그인이 안되어 있는데 글쓰기가 되는 문제.
- 같은 id 중복이어도 가입 가능한 문제
- 나머지 작은 이슈들

+++ src/main/java/study/community/controller/BoardController.java
@@ -10,6 +10,7 @@ import study.community.service.BoardService;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
+import java.util.Optional;
 
 @Controller
 public class BoardController {
@@ -26,28 +27,30 @@ public class BoardController {
         return "index";
     }
 
-    //List
+    //ListREAD
     @GetMapping("board/list")
     public String boardList(Model model){
         model.addAttribute("boardList",boardService.getBoardList());
         return "board/list";
     }
 
-    //Board
+    //READ
     @GetMapping("board/{idx}")
     public String board(@PathVariable("idx") Long idx, Model model){
-        model.addAttribute("board",boardService.getByIdx(idx));
+        model.addAttribute("board",boardService.getByIdx(idx).get());
 
         return "board/board";
     }
 
     //Write
     @GetMapping("board/new")
-    public String initCreateBoard(Board board){
-
-        return "board/new";
+    public String initCreateBoard(Board board, HttpServletRequest request) {
+        HttpSession session = request.getSession();
+        if (session.getAttribute("user") != null) {
+            return "board/new";
+        }
+        else return "redirect:/board/list";
     }
-    //Create
     @PostMapping("board/new")
     public String processingCreateBoard(Board board, HttpServletRequest request){
         User user = (User)request.getSession().getAttribute("user");
@@ -56,24 +59,34 @@ public class BoardController {
     }
     //Update
     @GetMapping("/board/{idx}/update")
-    public String initUpdateBoard(@PathVariable long idx, Model model,HttpServletRequest request){
-        Board board = boardService.getByIdx(idx);
-        model.addAttribute("board",board);
+    public String initUpdateBoard(@PathVariable Long idx, Model model,HttpServletRequest request){
+        HttpSession session = request.getSession();
+        //로그인 정보와 작성자가 일치하는지 확인
+        if(boardService.getByIdx(idx).get().getUser().getId().equals(((User)session.getAttribute("user")).getId())) {
+            Board board = boardService.getByIdx(idx).get();
+            model.addAttribute("board", board);
 
-        return "board/update";
+            return "board/update";
+        }
+        else return "redirect:/board/"+idx;
     }
     @PostMapping("board/{idx}/update")
-    public String processingUpdateBoard(@PathVariable long idx, Board board){
+    public String processingUpdateBoard(@PathVariable Long idx, Board board){
         board.setIdx(idx);
-        board.setUser(boardService.getByIdx(idx).getUser());
+        board.setUser(boardService.getByIdx(idx).get().getUser());
         boardService.updateBoard(board);
         return "redirect:/board/"+board.getIdx();
     }
 
     //Delete
     @GetMapping("board/{idx}/delete")
-    public String deleteBoard(@PathVariable long idx){
-        boardService.deleteBoardByIdx(idx);
-        return "redirect:/board/list";
+    public String deleteBoard(@PathVariable Long idx, HttpServletRequest request){
+        HttpSession session = request.getSession();
+        if(boardService.getByIdx(idx).get().getUser().getId().equals(((User)session.getAttribute("user")).getId())) {
+
+            boardService.deleteBoardByIdx(idx);
+            return "redirect:/board/list";
+        }
+        else return "redirect:/board/"+idx;
     }
 }
